Renaming map_addmodel.h to cnvmodel.h
