+++ codebase/superdarn/src.bin/tk/tool/map_addmodel.1.14/map_addmodel.c
@@ -33,7 +33,7 @@
 #include "shfconst.h" /* use the same constants as in fitting procedure */
 #include "igrfcall.h"
 #include "igrflib.h"
-#include "map_addmodel.h"
+#include "cnvmodel.h"
 #include "map_addhmb.h"
 #include "calc_bmag.h"
 #include "hlpstr.h"
